+++ opentracing-spring-cloud/pom.xml
@@ -107,6 +107,16 @@
       <artifactId>artemis-jms-client</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-artemis</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.activemq</groupId>
+      <artifactId>artemis-jms-server</artifactId>
+      <scope>test</scope>
+    </dependency>
 
     <dependency>
       <groupId>io.opentracing</groupId>
