+++ opentracing-spring-cloud/src/test/java/io/opentracing/contrib/spring/cloud/jms/JmsArtemisStarterTest.java
@@ -1,0 +1,108 @@
+package io.opentracing.contrib.spring.cloud.jms;
+
+import static org.awaitility.Awaitility.await;
+import static org.jgroups.util.Util.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
+import java.util.List;
+
+import javax.jms.ConnectionFactory;
+import javax.jms.Message;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.jms.DefaultJmsListenerContainerFactoryConfigurer;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.ResponseEntity;
+import org.springframework.jms.annotation.EnableJms;
+import org.springframework.jms.annotation.JmsListener;
+import org.springframework.jms.config.DefaultJmsListenerContainerFactory;
+import org.springframework.jms.config.JmsListenerContainerFactory;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import io.opentracing.contrib.spring.cloud.MockTracingConfiguration;
+import io.opentracing.contrib.spring.cloud.TestUtils;
+import io.opentracing.mock.MockSpan;
+import io.opentracing.mock.MockTracer;
+
+/**
+ * @author Pavol Loffay
+ */
+@SpringBootTest(
+        webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
+        classes = {
+                MockTracingConfiguration.class,
+                JmsArtemisStarterTest.JmsTestConfiguration.class,
+                MsgController.class})
+@RunWith(SpringJUnit4ClassRunner.class)
+public class JmsArtemisStarterTest {
+
+    @Configuration
+    @EnableJms
+    static class JmsTestConfiguration {
+
+        @Bean
+        public MsgListener msgListener() {
+            return new MsgListener();
+        }
+
+        @Bean
+        public JmsListenerContainerFactory<?> myFactory(ConnectionFactory connectionFactory,
+                                                        DefaultJmsListenerContainerFactoryConfigurer configurer) {
+            DefaultJmsListenerContainerFactory factory = new DefaultJmsListenerContainerFactory();
+            // This provides all boot's default to this factory, including the message converter
+            configurer.configure(factory, connectionFactory);
+            // You could still override some of Boot's default if necessary.
+            return factory;
+        }
+
+        public static class MsgListener {
+            private Message message;
+
+            public Message getMessage() {
+                return message;
+            }
+
+            @JmsListener(destination = "fooQueue", containerFactory = "myFactory")
+            public void processMessage(Message msg) throws Exception {
+                message = msg;
+            }
+        }
+    }
+
+    @Autowired
+    private MockTracer tracer;
+
+    @Autowired
+    private JmsTestConfiguration.MsgListener msgListener;
+
+    @Autowired
+    private TestRestTemplate restTemplate;
+
+    @Before
+    public void before() {
+        tracer.reset();
+    }
+
+    @Test
+    public void testListenerSpans() {
+        ResponseEntity<String> responseEntity = restTemplate.getForEntity("/hello", String.class);
+
+        await().until(() -> {
+            List<MockSpan> mockSpans = tracer.finishedSpans();
+            return (mockSpans.size() == 3);
+        });
+
+        assertNotNull(msgListener.getMessage());
+
+        assertEquals(200, responseEntity.getStatusCode().value());
+        List<MockSpan> spans = tracer.finishedSpans();
+        assertEquals(3, spans.size());  // it propagated over to @JmsListener
+        TestUtils.assertSameTraceId(spans);
+    }
+}
